BUG: fix SuiteSparse_long on windows
Fundamentally, SuiteSparse requires a 64 bit
integer to operate correctly. On GCC, this type
is provided by the "long" type definition.
However, although the "long" type happens to meet
this criteria, is not necesarrily the case that
a "long" is 64 bits because it is officially
defined to be between 32 and 64 bits.

Because MSVC implemented "long" differently from
GCC, although both complied with the specification,
MSVC failed when GCC did not.

To fix this error, the SWIG file was updated to use
the C99 int64_t, which is standardized at 64 bits.
A compatibility header was added for MSVC, which
does not fully support C99 (the compatibility
header can be removed sometime in the future).