Fix order-dependent SWIG includes
This is based on work done by @JohnDN90 in
https://github.com/JohnDN90/scikit-umfpack_workaround

It no longer requires modifying the suitesparse headers, and we just
```
%include <umfpack.h>
```
at the bottom of `umfpack.i`.

This is still a work-in-progress because the `*_free_numeric` and `*_free_symbolic`
functions now don't have correct bindings, so they are just commented out in `umfpack.py`.